moving authentication regulator to typescript
